viewer/tkinter-viewer/pyproject.toml
[project]
name = "tkinter-viewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]
